展会更新重构

+++ king-admin-java/src/main/java/com/oukingtim/service/impl/MgrExhibitionServiceImpl.java
@@ -12,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
+import org.springframework.util.StringUtils;
 
 import java.io.Serializable;
 import java.util.Date;
@@ -100,23 +101,20 @@ public class MgrExhibitionServiceImpl extends ServiceImpl<MgrExhibitionMapper, E
 
             //获取上传的展会详情
             ExhibitionDetail exhibitionDetail = exhibition.getExhibitionDetail();
-            if (exhibitionDetail != null) {
-                entity.setExhibitionId(exhibitionId);
+            if (exhibitionDetail != null && !StringUtils.isEmpty(exhibitionDetail.getId())) {
                 //删除展会详情
-                mgrExhibitionDetailMapper.delete(new EntityWrapper<>(entity));
+                /*entity.setExhibitionId(exhibitionId);
+                mgrExhibitionDetailMapper.delete(new EntityWrapper<>(entity));*/
 
                 exhibitionDetail.setExhibitionId(exhibitionId);
-                exhibitionDetail.setCreateTime(new Date());
                 exhibitionDetail.setUpdateTime(new Date());
-                //插入展会详情
-                mgrExhibitionDetailMapper.insert(exhibitionDetail);
+                //更新展会详情
+                mgrExhibitionDetailMapper.updateById(exhibitionDetail);
                 File file = new File();
                 file.setTypeId(exhibitionDetail.getId());
                 //删除展会详情相关的文件(图片)
                 mgrFileMapper.delete(new EntityWrapper<>(file));
                 if (!CollectionUtils.isEmpty(exhibitionDetail.getFiles())) {
-
-
                     //更新展会详情文件
                     for (File fileEntity : exhibitionDetail.getFiles()) {
                         fileEntity.setTypeId(exhibitionDetail.getId());
